# Conflicts:
#	LiuMa-backend/src/main/java/com/autotest/LiuMa/common/utils/EmailUtils.java
#	LiuMa-backend/src/main/java/com/autotest/LiuMa/database/domain/Case.java
#	LiuMa-backend/src/main/java/com/autotest/LiuMa/database/domain/Report.java
#	LiuMa-backend/src/main/java/com/autotest/LiuMa/request/CaseApiRequest.java
#	LiuMa-backend/src/main/java/com/autotest/LiuMa/request/CaseRequest.java
#	LiuMa-backend/src/main/java/com/autotest/LiuMa/request/RunRequest.java
#	LiuMa-backend/src/main/java/com/autotest/LiuMa/service/OpenApiService.java
#	LiuMa-backend/src/main/resources/application.properties
#	LiuMa-backend/src/main/resources/generatorConfig.xml
#	LiuMa-frontend/config/index.js
#	LiuMa-frontend/package-lock.json
#	LiuMa-frontend/src/views/baseCenter/operationEdit.vue
#	LiuMa-frontend/src/views/caseCenter/apiCaseEdit.vue
#	LiuMa-frontend/src/views/caseCenter/common/case/controller.vue
#	LiuMa-frontend/src/views/common/components/navcon.vue

+++ LiuMa-backend/src/main/java/com/autotest/LiuMa/service/ModuleService.java
@@ -1,6 +1,7 @@
 package com.autotest.LiuMa.service;
 
-import com.autotest.LiuMa.common.exception.DuplicateContentException;
+import com.autotest.LiuMa.common.exception.DuplicateException;
+import com.autotest.LiuMa.common.exception.LMException;
 import com.autotest.LiuMa.database.mapper.ModuleMapper;
 import com.autotest.LiuMa.dto.ModuleDTO;
 import org.springframework.stereotype.Service;
@@ -19,9 +20,10 @@ public class ModuleService {
     private ModuleMapper moduleMapper;
 
     public ModuleDTO save(ModuleDTO module) {
-        ModuleDTO oldModule = moduleMapper.getModuleByParentAndName(module.getModuleType(), module.getName(), module.getParentId());
+        ModuleDTO oldModule = moduleMapper.getModuleByParentAndName(module.getModuleType(),
+                module.getName(), module.getParentId(), module.getProjectId());
         if(oldModule != null){
-            throw new DuplicateContentException("当前父模块下已有重名模块");
+            throw new DuplicateException("当前父模块下已有重名模块");
         }
         if(module.getId() == null){
             //新增模块
@@ -39,12 +41,11 @@ public class ModuleService {
     }
 
     public void delete(ModuleDTO module) {
-        moduleMapper.deleteModule(module.getModuleType(), module.getId());
-        if(module.getChildren() != null) {
-            for (ModuleDTO moduleDTO : module.getChildren()) {
-                this.delete(moduleDTO);
-            }
+        Integer count = moduleMapper.getModuleDataById(module.getModuleType(), module.getId());
+        if(count>0){
+            throw new LMException("当前模块下已有相关数据，无法删除！");
         }
+        moduleMapper.deleteModule(module.getModuleType(), module.getId());
     }
 
     public List<ModuleDTO> getModuleList(String moduleType, String projectId){
